+++ scripts/ci/notcurses-install-deps.sh
@@ -2,7 +2,6 @@
 set -ex
 sudo apt -q update
 sudo apt install -y \
-            libkf5windowsystem5 \
             libqt5core5a \
             libqt5gui5 \
             libqt5network5 \
@@ -22,9 +21,8 @@ sudo apt install -y \
             libfontconfig1 \
             libfreetype6 \
             libharfbuzz0b \
-            libkf5windowsystem5 \
             libqt5core5a \
             libqt5gui5 \
             libqt5network5 \
-						libqt5multimedia5 \
+            libqt5multimedia5 \
             libyaml-cpp0.6